pyproject.toml
[project]
name = "telegram-news"
version = "0.1.0"
description = "Telegram Crypto News Aggregator"
requires-python = ">=3.11"
dependencies = [
    "telethon>=1.37",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "trafilatura>=2.0",
    "playwright>=1.49",
]

[project.optional-dependencies]
dev = [
    "ruff",
]
